refactor(tools): extract extension_from_path to mod.rs
Both manifest.rs and context.rs defined identical private
`extension_from_path` helpers — a drift hazard any time the guard logic
(rejecting dotfiles like .gitignore) needs updating. Hoist the shared
helper to src/tools/mod.rs as pub(crate) and consolidate the test set
there. Callers in manifest.rs and context.rs import it from the parent
module.

The pub(crate) visibility keeps this an internal orchestration detail:
downstream MCP clients don't need access to the helper, and keeping it
non-public means we can change its behavior (e.g., handle multi-dot
extensions differently) without a SemVer bump.
