+++ src/tools/manifest.rs
@@ -8,6 +8,7 @@ use crate::git::diff::ChangeType;
 use crate::git::generated::GeneratedFileDetector;
 use crate::git::reader::RepoReader;
 use crate::pagination::{CURSOR_VERSION, PaginationCursor, PaginationInfo, encode_cursor};
+use crate::tools::extension_from_path;
 use crate::tools::size;
 use crate::tools::types::{
     FunctionChange, FunctionChangeType, ImportChange, ManifestFileEntry, ManifestMetadata,
@@ -117,13 +118,6 @@ pub fn diff_imports(base_imports: &[String], head_imports: &[String]) -> ImportC
     removed.sort();
 
     ImportChange { added, removed }
-}
-
-fn extension_from_path(path: &str) -> &str {
-    path.rsplit('.')
-        .next()
-        .filter(|ext| path.len() > ext.len() + 1)
-        .unwrap_or("")
 }
 
 fn matches_glob_pattern(path: &str, pattern: &str) -> bool {
@@ -1851,33 +1845,6 @@ mod tests {
         let changes = diff_functions(&base, &head);
         assert_eq!(changes[0].name, "alpha");
         assert_eq!(changes[1].name, "zebra");
-    }
-
-    // --- extension_from_path tests ---
-
-    #[test]
-    fn it_extracts_extension_from_normal_path() {
-        assert_eq!(extension_from_path("src/main.rs"), "rs");
-    }
-
-    #[test]
-    fn it_returns_empty_for_dotfile() {
-        assert_eq!(extension_from_path(".gitignore"), "");
-    }
-
-    #[test]
-    fn it_returns_empty_for_no_extension() {
-        assert_eq!(extension_from_path("Makefile"), "");
-    }
-
-    #[test]
-    fn it_extracts_last_extension_from_multiple_dots() {
-        assert_eq!(extension_from_path("archive.tar.gz"), "gz");
-    }
-
-    #[test]
-    fn it_returns_empty_for_empty_string() {
-        assert_eq!(extension_from_path(""), "");
     }
 
     // --- matches_glob_pattern tests ---
